move diagnostics to new file
